rag-pareto-frontiers/requirements.txt
numpy
pandas
pyyaml
tqdm
matplotlib
scikit-learn
rank-bm25
sentence-transformers
faiss-cpu
rapidfuzz
requests
pytest
